+++ epub3/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.2.2
+Fix path in ContentReader.
+Add Book.readBytes, Book.readString.
+Replace `part of` to import/export.
+
+# 0.2.0
+Rename to epub3 for search order of pub.dev
+
 # 0.1.0
 Add Writer with content and pass epubcheck.
 Change all interface.